third_party/blink/web_tests/wpt_internal/prerender/restriction-web-serial.https.html

<!DOCTYPE html>
<!--
This file cannot be upstreamed to WPT until:
* The defer expectation is replaced with rejecting the promise with
  “SecurityError” to match the specification at:
  https://wicg.github.io/nav-speculation/prerendering.html#patch-serial
-->
<title>Access to the Web Serial API is deferred</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('test-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  const url = `resources/web-serial.https.html`;
  window.open(url, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {event: 'started waiting navigator.serial.getPorts',  prerendering: true},
    {event: 'prerendering change',                        prerendering: false},
    {event: 'finished waiting navigator.serial.getPorts', prerendering: false},
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event[${i}]`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering[${i}]`);
  }
  bc.close();
}, `the access to the Web Serial API should be deferred until the prerendered
    page is activated`);

</script>
</body>
